requirements/dev-requirements.txt
black==23.1.0
django-debug-toolbar==3.8.1 
django-dump-load-utf8==0.0.4
click==8.1.3
colorama==0.4.6

mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
tomli==2.0.1
